Improve support of large log file on 32-bit host (#308)
* Better support large log file when -D_FILE_OFFSET_BITS=64 on 32-bit host

* Use off_t to define log file size uniformly

Cherry-pick: 39e5c724c8f34e084930b9b668d1caaa784aaf0c

+++ muduo/base/AsyncLogging.cc
@@ -7,7 +7,7 @@
 using namespace muduo;
 
 AsyncLogging::AsyncLogging(const string& basename,
-                           size_t rollSize,
+                           off_t rollSize,
                            int flushInterval)
   : flushInterval_(flushInterval),
     running_(false),
